requirements.txt
absl-py==0.12.0
aiohttp==3.7.4
antlr4-python3-runtime==4.8
astroid==2.5
async-timeout==3.0.1
attrs==20.3.0
blinker==1.4
brotlipy==0.7.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
colorlog==5.0.1
cryptography==3.4.7
cycler==0.10.0
fsspec==0.9.0
future==0.18.2
google-auth==1.28.0
google-auth-oauthlib==0.4.1
grpcio==1.33.2
hydra-colorlog==1.0.1
hydra-core==1.0.6
idna==2.10
importlib-metadata==3.10.0
importlib-resources==5.1.2
isort==5.8.0
joblib==1.0.1
kiwisolver==1.3.1
kornia==0.5.1
lazy-object-proxy==1.6.0
Markdown==3.3.4
matplotlib==3.4.1
mccabe==0.6.1
multidict==5.1.0
numpy==1.19.2
oauthlib==3.0.1
olefile==0.46
omegaconf==2.0.6
packaging==20.9
Pillow==8.2.0
pip==21.0.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.20
PyJWT==2.0.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytorch-lightning==1.2.7
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
setuptools==52.0.0.post20210125
six==1.15.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.2
threadpoolctl==2.1.0
toml==0.10.2
torch==1.7.1
torchmetrics==0.2.0
torchvision==0.8.2
tqdm==4.60.0
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
wheel==0.36.2
wrapt==1.12.1
yarl==1.6.3
zipp==3.4.1